refactor: tear down `ContainerWithGridSelection` in alldayGridRow
